Match evergreen browsers with evergreen drivers
It's my understanding that, previously, browsers had to be updated manually; it seems today's browsers update *automatically*. To match this, this commit makes use of a WebDriverManager that automatically detects and downloads the correct browser driver for the currently installed browser.

This should meet everyone where they are: CI boxes that have old browsers or new people who want the example to work out of the box.